da: move da client to new file; fix devnet version
