+++ ops-bedrock/docker-compose.yml
@@ -21,7 +21,7 @@ services:
     image: us-docker.pkg.dev/oplabs-tools-artifacts/images/op-stack-go:devnet
     entrypoint: ["echo", "build complete"]
   da:
-    image: opcelestia/local-celestia-devnet:v0.12.0-rc0
+    image: opcelestia/local-celestia-devnet:v0.12.1-rc0
     ports:
       - "26650:26650"
       - "26658:26658"
